Move tvOS/iOS specific functions to separate files
